Create test body classes
