+++ test/timeline/drag/global/global_record_timeline_drag_test.dart
@@ -1,17 +1,17 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/src/timeline/timeline.dart';
 
-import '../timeline_drag_test_bodies.dart';
+import '../act_drag_timeline_test_bodies.dart';
 
 void main() {
   globalTimelineMode = TimelineMode.record;
   testWidgets('Global: record, turn off during test', (tester) async {
-    await recordTurnOffDuringTestDrag(tester, isGlobal: true);
+    await ActDragTimelineTestBodies.recordTurnOff(tester, isGlobal: true);
   });
   testWidgets('Global: record - without error', (tester) async {
-    await recordTurnOffDuringTestDrag(tester, isGlobal: true);
+    await ActDragTimelineTestBodies.recordNoError(tester, isGlobal: true);
   });
   test('Global: record, with error, prints timeline and html', () async {
-    await recordWithErrorPrintsHTML(isGlobal: true);
+    await ActDragTimelineTestBodies.recordWithErrorPrintsHTML(isGlobal: true);
   });
 }
