--- test/timeline/drag/timeline_drag_test.dart
@@ -1,194 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter_test/flutter_test.dart';
-import 'package:spot/spot.dart';
-import 'package:test_process/test_process.dart';
-import '../timeline_test_shared.dart' as shared;
-import 'timeline_drag_test_bodies.dart';
-
-final _secondItemSelector = spotText('Item at index: 27', exact: true);
-const _passingDragAmount = 23;
-const _passingOffset = Offset(0, -2300);
-
-const _header = '==================== Timeline Event ====================';
-
-String _testAsString({
-  required String title,
-  required TimelineMode timelineMode,
-  required int drags,
-}) {
-  final widgetPart =
-      File('test/timeline/drag/drag_until_visible_test_widget.dart')
-          .readAsStringSync();
-  return '''
-import 'package:flutter_test/flutter_test.dart';
-import 'package:spot/spot.dart';
-import 'package:spot/src/timeline/timeline.dart';\n
-$widgetPart\n
-void main() async {
-  testWidgets("$title", (WidgetTester tester) async {
-  ${shared.localTimelineInitiator(timelineMode)};
-    await tester.pumpWidget(const DragUntilVisibleTestWidget());
-      final firstItem = spotText('Item at index: 3', exact: true)..existsOnce();
-      final secondItem = spotText('Item at index: 27', exact: true)
-        ..doesNotExist();
-      await act.dragUntilVisible(
-        dragStart: firstItem,
-        dragTarget: secondItem,
-        maxIteration: $drags,
-        moveStep: const Offset(0, -100),
-      );
-      secondItem.existsOnce();
-  });
-}
-''';
-}
-
-void main() {
-  group('Drag Timeline Test', () {
-    group('Without error', () {
-      testWidgets('Local: record, without error', (tester) async {
-        await liveTimelineWithoutErrorsDrag(tester);
-      });
-      testWidgets('Local: off does not record', (tester) async {
-        await offTimelineWithoutErrorsDrag(tester);
-      });
-      testWidgets('Local: record, turn off during test', (tester) async {
-        await recordTurnOffDuringTestDrag(tester);
-      });
-      testWidgets('Local: record - without error', (tester) async {
-        await recordNoErrorsDrag(tester);
-      });
-    });
-    group('Teardown test', () {
-      test('Local: record, with error, prints timeline and html', () async {
-        await recordWithErrorPrintsHTML();
-      });
-      test('Local: live - without error, prints HTML', () async {
-        await liveWithoutErrorPrintsHTML();
-      });
-      test('Local: live - with error, prints HTML, no duplicates', () async {
-        await liveWitErrorPrintsHTMLNoDuplicates();
-      });
-      // test('Live timeline - with error, no duplicates, prints HTML', () async {
-      //   final tempDir = Directory.systemTemp.createTempSync();
-      //   final tempTestFile = File('${tempDir.path}/temp_test.dart');
-      //   await tempTestFile.writeAsString(
-      //     _testAsString(
-      //       title: 'Live timeline - with error, no duplicates, prints HTML',
-      //       timelineMode: TimelineMode.live,
-      //       shouldFail: true,
-      //     ),
-      //   );
-      //   final testProcess =
-      //       await TestProcess.start('flutter', ['test', tempTestFile.path]);
-      //   final stdoutBuffer = StringBuffer();
-      //   bool write = false;
-      //   await for (final line in testProcess.stdoutStream()) {
-      //     if (line.isEmpty) continue;
-      //     if (!write) {
-      //       if (line == _header) {
-      //         write = true;
-      //       }
-      //       if (write) {
-      //         stdoutBuffer.writeln(line);
-      //       } // Error does not happen
-      //       await testProcess.shouldExit(1);
-      //       tempDir.deleteSync(recursive: true);
-      //       final stdout = stdoutBuffer.toString();
-      //       final timeline = stdout.split('\n');
-      //       // Does not start with 'Timeline', this only happens on error
-      //       expect(timeline.first, _header);
-      //       expect(
-      //         timeline.second,
-      //         'Event: Tap Icon Widget with icon: "IconData(U+0E047)"',
-      //       );
-      //       expect(
-      //         timeline[2].startsWith('Caller: at main.<fn> file:///'),
-      //         isTrue,
-      //       );
-      //       expect(
-      //         timeline[3].startsWith(
-      //           'Screenshot: file:///',
-      //         ),
-      //         isTrue,
-      //       );
-      //       expect(
-      //         timeline[4].startsWith(
-      //           'Timestamp:',
-      //         ),
-      //         isTrue,
-      //       );
-      //       expect(
-      //         timeline[5],
-      //         _separator,
-      //       );
-      //       final htmlLine = timeline
-      //           .firstWhere((line) => line.startsWith('View time line here:'));
-      //       expect(
-      //         htmlLine.endsWith(
-      //           'live_timeline_-_with_error,_no_duplicates,_prints_html.html',
-      //         ),
-      //         isTrue,
-      //       );
-      //     }
-      //   }
-      // });
-    });
-  });
-}
-
-void _testTimeLineContent({
-  required String output,
-  required Offset totalExpectedOffset,
-  required int drags,
-  bool runInTestProcess = false,
-}) {
-  final isGoingDown = totalExpectedOffset.dy < 0;
-  final findsWidget = drags == _passingDragAmount;
-  final eventLines =
-      output.split('\n').where((line) => line.startsWith('Event:'));
-  final startEvent = _replaceOffsetWithDxDy(eventLines.first);
-  final endEvent =
-      'Event: Target ${_secondItemSelector.toStringBreadcrumb()} ${findsWidget ? 'found' : 'not found'} after $drags drags. Total dragged offset: $totalExpectedOffset';
-
-  for (int i = 0; i < 2; i++) {
-    expect(
-      RegExp(_header).allMatches(output).length,
-      2,
-    );
-    if (i == 0) {
-      expect(
-        startEvent,
-        'Event: Scrolling ${isGoingDown ? 'downwards' : 'upwards'} from Offset(dx,dy) in order to find ${_secondItemSelector.toStringBreadcrumb()}.',
-      );
-    } else {
-      expect(
-        eventLines.last,
-        endEvent,
-      );
-    }
-    expect(
-      RegExp('Caller: at ${runInTestProcess ? 'main' : '_testBody'}')
-          .allMatches(output)
-          .length,
-      2,
-    );
-    expect(
-      RegExp('Screenshot: file:').allMatches(output).length,
-      2,
-    );
-    expect(
-      RegExp('Timestamp: ').allMatches(output).length,
-      2,
-    );
-  }
-}
-
-String _replaceOffsetWithDxDy(String originalString) {
-  // Regular expression to match 'Offset(any values)'
-  final RegExp offsetPattern = RegExp(r'Offset\([^)]*\)');
-
-  // Replace all matches with 'Offset(dx,dy)'
-  return originalString.replaceAll(offsetPattern, 'Offset(dx,dy)');
-}
