+++ test/timeline/tap/global/global_live_timeline_tap_test.dart
@@ -1,27 +1,27 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/src/timeline/timeline.dart';
-import '../timeline_tap_test_bodies.dart' as body;
+import '../act_tap_timeline_test_bodies.dart';
 
 void main() {
   globalTimelineMode = TimelineMode.live;
   testWidgets('Global: record, without error', (tester) async {
-    await body.liveWithoutError(
+    await ActTapTimelineTestBodies.liveWithoutError(
       tester: tester,
       isGlobalMode: true,
     );
   });
   test('Global: live - without error, prints HTML', () async {
-    await body.liveWithoutErrorPrintsHtml(
+    await ActTapTimelineTestBodies.liveWithoutErrorPrintsHtml(
       isGlobalMode: true,
     );
   });
   test('Global: live - with error, no duplicates, prints HTML', () async {
-    await body.liveWithErrorNoDuplicatesPrintsHtml(
+    await ActTapTimelineTestBodies.liveWithErrorNoDuplicatesPrintsHtml(
       isGlobalMode: true,
     );
   });
   testWidgets('Global: live, turn off during test', (tester) async {
-    await body.liveTurnOffDuringTest(
+    await ActTapTimelineTestBodies.liveTurnOffDuringTest(
       isGlobalMode: true,
       tester: tester,
     );
