+++ test/timeline/tap/global/global_off_timeline_tap_test.dart
@@ -1,11 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/src/timeline/timeline.dart';
-import '../timeline_tap_test_bodies.dart' as body;
+import '../act_tap_timeline_test_bodies.dart';
 
 void main() {
   globalTimelineMode = TimelineMode.off;
   testWidgets('Global: off, without error', (tester) async {
-    await body.offWithoutError(
+    await ActTapTimelineTestBodies.offWithoutError(
       tester: tester,
       isGlobalMode: true,
     );
