+++ test/timeline/tap/global/global_record_timeline_tap_test.dart
@@ -1,16 +1,16 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/src/timeline/timeline.dart';
-import '../timeline_tap_test_bodies.dart' as body;
+import '../act_tap_timeline_test_bodies.dart';
 
 void main() {
   globalTimelineMode = TimelineMode.record;
   testWidgets('Global: record, without error', (tester) async {
-    await body.recordWithoutError(
+    await ActTapTimelineTestBodies.recordWithoutError(
       tester: tester,
       isGlobalMode: true,
     );
   });
   test('Global: record, with error', () async {
-    await body.recordWithError(isGlobalMode: true);
+    await ActTapTimelineTestBodies.recordWithError(isGlobalMode: true);
   });
 }
